move NotificationIcon to notifications folder
